+++ tests/Fixtures/TestPanelProvider.php
@@ -4,8 +4,17 @@ declare(strict_types=1);
 
 namespace Albertoarena\FilamentEventSourcing\Tests\Fixtures;
 
+use Albertoarena\FilamentEventSourcing\FilamentEventSourcingPlugin;
+use Filament\Http\Middleware\Authenticate;
+use Filament\Http\Middleware\DisableBladeIconComponents;
+use Filament\Http\Middleware\DispatchServingFilamentEvent;
 use Filament\Panel;
 use Filament\PanelProvider;
+use Illuminate\Cookie\Middleware\AddQueuedCookiesToResponse;
+use Illuminate\Cookie\Middleware\EncryptCookies;
+use Illuminate\Routing\Middleware\SubstituteBindings;
+use Illuminate\Session\Middleware\StartSession;
+use Illuminate\View\Middleware\ShareErrorsFromSession;
 
 final class TestPanelProvider extends PanelProvider
 {
@@ -15,6 +24,22 @@ final class TestPanelProvider extends PanelProvider
             ->default()
             ->id('admin')
             ->path('admin')
-            ->authGuard('web');
+            ->authGuard('web')
+            ->middleware([
+                EncryptCookies::class,
+                AddQueuedCookiesToResponse::class,
+                StartSession::class,
+                ShareErrorsFromSession::class,
+                SubstituteBindings::class,
+                DisableBladeIconComponents::class,
+                DispatchServingFilamentEvent::class,
+            ], isPersistent: true)
+            ->authMiddleware([
+                Authenticate::class,
+            ])
+            ->plugin(
+                FilamentEventSourcingPlugin::make()
+                    ->storedEventsResource()
+            );
     }
 }